requirements.txt
aiohttp
beautifulsoup4
celery>=5.3
django
django-celery-results>=2.5
hdbscan
numpy
pandas
pgvector>=0.3
psycopg[binary]
redis>=5.0
requests
scikit-learn
sentence-transformers
tqdm
umap-learn
